refactor common.h into multiple .h and .cpp files
