+++ src/bedrock/monitoring.cpp
@@ -8,11 +8,16 @@
 #include <pthread.h>
 #include <unistd.h>
 #include <memory>
+#include "spdlog/spdlog.h"
 #include "communication.pb.h"
 #include "zmq/socket_cache.h"
 #include "zmq/zmq_util.h"
 #include "utils/consistent_hash_map.hpp"
 #include "common.h"
+#include "threads.h"
+#include "requests.h"
+#include "hashers.h"
+#include "hash_ring.h"
 #include "yaml-cpp/yaml.h"
 
 // the default number of nodes to add concurrently for storage
