+++ src/bedrock/routing.cpp
@@ -8,11 +8,15 @@
 #include <pthread.h>
 #include <unistd.h>
 #include <memory>
+#include "spdlog/spdlog.h"
 #include "communication.pb.h"
 #include "zmq/socket_cache.h"
 #include "zmq/zmq_util.h"
 #include "utils/consistent_hash_map.hpp"
 #include "common.h"
+#include "threads.h"
+#include "hashers.h"
+#include "hash_ring.h"
 #include "yaml-cpp/yaml.h"
 
 using namespace std;
