+++ src/bedrock/user.cpp
@@ -8,10 +8,15 @@
 #include <unistd.h>
 #include <memory>
 #include <unordered_set>
+#include "spdlog/spdlog.h"
 #include "communication.pb.h"
 #include "zmq/socket_cache.h"
 #include "zmq/zmq_util.h"
 #include "common.h"
+#include "threads.h"
+#include "requests.h"
+#include "hashers.h"
+#include "hash_ring.h"
 #include "yaml-cpp/yaml.h"
 
 using namespace std;
